Add files via upload
The Titanic dataset project is a machine learning model that predicts the survival of passengers on the Titanic based on a set of input features such as age, gender, passenger class, and fare. The dataset includes information on over 891 passengers , including whether they survived or not. The model is trained on this historical data, and then used to make predictions on new, unseen data. The goal of the project is to create a model that is accurate in predicting survival, and to gain insights into the factors that influenced survival rates on the Titanic.